new: hook to get token balance of any account
